extract ViewerBody and ViewerMenu components
